requirements.txt
Flask
Flask-WTF
WTForms
requests
python-dotenv
sounddevice
numpy
soundfile
openai
